+++ config/config_internal_other_platforms_test.go
@@ -1,0 +1,13 @@
+//go:build !darwin || !arm64
+
+package config
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestDefaultInstallDirWithNotMacOSARM64(t *testing.T) {
+	assert.Equal(t, "/usr/local/bin", defaultInstallDir())
+}